Add permission dan tambah divisi di form pengajuan bahan

+++ resources/views/livewire/pengajuan-table.blade.php
@@ -42,11 +42,11 @@
                     @include('livewire.dataperpage')
                 </li>
                 <li class="m-1">
-                    {{-- @can('tambah-projek') --}}
+                    @can('tambah-pengajuan')
                         <a href="{{ route('pengajuans.create') }}" class="mt-2 block w-fit rounded-md py-1.5 px-2 bg-indigo-600 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">
                             Tambah
                         </a>
-                    {{-- @endcan --}}
+                    @endcan
                 </li>
             </ul>
         </div>
@@ -99,16 +99,16 @@
                                         </a>
                                     @endif --}}
                                     @if($pengajuan->status !== 'Konfirmasi')
-                                        {{-- @can('edit-projek') --}}
+                                        @can('edit-pengajuan')
                                             <a href="{{ route('pengajuans.edit', $pengajuan->id) }}" class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-yellow-600 hover:border-yellow-600 focus:text-white focus:bg-yellow-600 focus:border-yellow-600 active:border-yellow-600 active:text-white active:bg-yellow-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none">
                                                 <svg class="w-[16px] h-[16px] text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
                                                 <path stroke="currentColor" stroke-width="2" d="m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z"/>
                                                 </svg>
                                             </a>
-                                        {{-- @endcan --}}
+                                        @endcan
                                     @endif
                                     @if ($pengajuan->status === 'Konfirmasi')
-                                        {{-- @can('hapus-projek') --}}
+                                        @can('hapus-pengajuan')
                                             <button wire:click="deletePengajuans({{ $pengajuan->id }})"
                                                 data-modal-target="deleteproduksi-modal" data-modal-toggle="deleteproduksi-modal"
                                                 class="rounded-md border border-slate-300 py-1 px-2 text-center text-xs transition-all shadow-sm hover:shadow-lg text-slate-600 hover:text-white hover:bg-red-600 hover:border-red-600 focus:text-white focus:bg-red-600 focus:border-red-600 active:border-red-600 active:text-white active:bg-red-600 disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
@@ -121,7 +121,7 @@
                                                         d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z" />
                                                 </svg>
                                             </button>
-                                        {{-- @endcan --}}
+                                        @endcan
                                     @endif
                                 </div>
                             </td>
